# Conflicts:
#	core/src/main/scala/kafka/security/authorizer/AclEntry.scala
#	core/src/main/scala/kafka/server/BrokerServer.scala
#	core/src/main/scala/kafka/server/ConfigHandler.scala
#	core/src/main/scala/kafka/server/KafkaConfig.scala
#	core/src/test/scala/unit/kafka/admin/ConfigCommandTest.scala
#	core/src/test/scala/unit/kafka/server/DynamicConfigChangeTest.scala
#	group-coordinator/src/main/java/org/apache/kafka/coordinator/group/GroupCoordinatorService.java
#	group-coordinator/src/main/java/org/apache/kafka/coordinator/group/GroupCoordinatorShard.java
#	group-coordinator/src/main/java/org/apache/kafka/coordinator/group/GroupMetadataManager.java
#	group-coordinator/src/main/java/org/apache/kafka/coordinator/group/runtime/CoordinatorRuntime.java
#	group-coordinator/src/test/java/org/apache/kafka/coordinator/group/runtime/CoordinatorRuntimeTest.java

+++ clients/src/main/java/org/apache/kafka/common/requests/AlterReplicaLogDirsRequest.java
@@ -83,7 +83,7 @@ public class AlterReplicaLogDirsRequest extends AbstractRequest {
         data.dirs().forEach(alterDir ->
             alterDir.topics().forEach(topic ->
                 topic.partitions().forEach(partition ->
-                    result.put(new TopicPartition(topic.name(), partition.intValue()), alterDir.path())))
+                    result.put(new TopicPartition(topic.name(), partition), alterDir.path())))
         );
         return result;
     }
